Project Euler 001
The most computationally efficient way to solve the problem of summing
all numbers 1-999 that are divisible by either 3 or 5. Simply iterating
the numbers and testing for divisibility requires more computing time
since we'd be testing for divisibility a minimum of 997 times (3 to
999). Instead, we factor out a 3, 5, and 15 from the three for-loops
respectively and sum them individually. At the end, we multiply each of
the sums by the number we factored out and add them together to get the
total.