Project Euler 003
The program tests all odd numbers between 3 and the integer value of the
square root of the number plus one for factorability of our large
number. Then it tests factors of the large number for primeness. Updates
can make it more efficient by using a while loop and a number%=i; line
to shrink the large number each time a factor is discovered to make the
program run more quickly.